+++ scripts/legacyMacros.js
@@ -553,6 +553,7 @@ export {surpriseAttack} from './macros/2014/monsterFeatures/generic/surpriseAtta
 export {swarmDamage} from './macros/2014/monsterFeatures/generic/swarmDamage.js';
 export {touchDamage} from './macros/2014/monsterFeatures/generic/touchDamage.js';
 export {undeadFortitude} from './macros/2014/monsterFeatures/generic/undeadFortitude.js';
+export {autoProne} from './macros/2014/monsterFeatures/generic/autoProne.js';
 // Actions
 export {checkCover} from './macros/2014/actions/checkCover.js';
 export {dash} from './macros/2014/actions/dash.js';