+++ scripts/macros/2024/monsterFeatures/zhentilarSoldier/zhentilarSoldierKnockDown.js
@@ -1,0 +1,8 @@
+export let zhentilarSoldierKnockDown = {
+    name: 'Knock Down',
+    version: '1.3.132',
+    rules: 'modern',
+    monsters: [
+        'Zhentilar Soldier'
+    ]
+};